interate compose method into createResolver instead of a separate helper
